pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "git_calendar"
authors = [{name = "Richard Darst"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    #"yaml2ics",  # now vendored
    #"yaml2ics @ https://github.com/scientific-python/yaml2ics/archive/refs/heads/main.zip",
    "python-dateutil",
    "jinja2",
    "mutt-ics",
    "pyyaml",
    "markdown-it-py",
    # There is some new unreleased ics-py feature needed.  As of 2025,
    # last release was in 2022.
    "ics @ https://github.com/ics-py/ics-py/archive/3b7d072.zip",  # main probably works too
]

[project.scripts]
git-calendar = "git_calendar.build:main"

[project.urls]
Home = "https://github.com/coderefinery/git-calendar"
